refactor/api wip
