+++ fea-rs/src/lib.rs
@@ -3,7 +3,6 @@
 mod compile;
 mod diagnostic;
 mod parse;
-mod source;
 mod token_tree;
 mod types;
 pub mod util;
